clk: brcmstb: introduce brcm,mux-clock
LCCF offers the mux-clock for muxes.  However, by default the
MUX will only select a frequency that is lower or equal to
the one requested.  This was a problem because certain situations
require the closest frequency.

This commit overcomes this problem by introducing the
"brcm,mux-clk" which by default will select the closest frequency
(even if it is above the requested).

refs #SWLINUX-4319

Signed-off-by: Jim Quinlan <james.quinlan@broadcom.com>