soc: brcmstb: pm: Cache DDR_PHY_STATUS_REG access before suspend
Make an access to the DDR_PHY_STATUS_REG base register address before
suspending in order to force the TLB to cache the translation to that
register.

Without doing this, during suspend, we could be seeing huge latencies
(up to 200ms) while accessing the register because the translation would
not be cached. This would require the TLB to go fetch the entry from
DRAM, which was in self-refresh.

refs #SWLINUX-4427

Signed-off-by: Florian Fainelli <florian.fainelli@broadcom.com>