hid: atv-remote: fix buffer overflow
In atvr_raw_event while handling a keycode passed in an audio packet,
we allocate a local key_data buffer which is passed to hid-core for
processing. hid_report_raw_event expects the size of the provided
buffer to match the report size and if smaller, it zeroes out the
remaining bytes in the buffer. In this case the buffer provided is
smaller than what is expected which makes the zeroing logic write
past the buffer length.

The issue is fixed by correctly allocating the buffer based on the
report size.

refs #SWANDROID-4226

Signed-off-by: Danesh Petigara <danesh.petigara@broadcom.com>