init: add wait for Android block device
The treble early mount feature requires the Android eMMC block
device that contains it's partitions to be available as soon as
init runs as the mounting of those partitions is done at a very
early stage within init.

This change ensures the Android block device is available
before init is launched to prevent early mount failures.

refs #SWANDROID-4568

Change-Id: Iceca53b2bba7bc20fe7f585bf84affad2a8b67e4
Signed-off-by: Danesh Petigara <danesh.petigara@broadcom.com>